Add files via upload
I successfully downloaded the data and created the function for once but I had some trouble creating the plot.  I tried using a session example as a guide but I still could not seem to accurately graph the data.  Time permitted, I would love to quickly run through this in class.